+++ core/core.go
@@ -14,6 +14,7 @@ import (
 	"chain/core/fetch"
 	"chain/core/leader"
 	"chain/core/mockhsm"
+	"chain/core/rpc"
 	"chain/core/txdb"
 	"chain/crypto/ed25519"
 	"chain/crypto/ed25519/chainkd"
@@ -22,7 +23,6 @@ import (
 	"chain/errors"
 	"chain/log"
 	"chain/net/http/httpjson"
-	"chain/net/rpc"
 	"chain/protocol"
 	"chain/protocol/bc"
 	"chain/protocol/state"
